backend/requirements.txt
Django==5.0.7
psycopg==3.2.1
psycopg-binary==3.2.1
djangorestframework==3.15.2
markdown==3.6
django-filter==24.2
python-dotenv==1.0.1
cloudinary==1.40.0
drf-yasg==1.21.7
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
